Changed to Jtable and finished home page
Added Jtable with the extension Jtable search bar. The search bar has
been edited as so that the reset button only is shown if the edit and
delete functions are also shown. Showing the reset button without the
other options caused extra colours that were not accounted for by the
table implementation to be created.